+++ pkg/client/cli/cmd_helm.go
@@ -8,6 +8,7 @@ import (
 	"github.com/spf13/pflag"
 
 	"github.com/telepresenceio/telepresence/rpc/v2/connector"
+	"github.com/telepresenceio/telepresence/rpc/v2/manager"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/ann"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/util"
 	"github.com/telepresenceio/telepresence/v2/pkg/errcat"
@@ -27,10 +28,15 @@ type helmArgs struct {
 	valuePairs []string
 	request    *connector.ConnectRequest
 	kubeFlags  *pflag.FlagSet
+	mode       manager.Mode
 }
 
 func helmInstallCommand() *cobra.Command {
-	var upgrade bool
+	var (
+		upgrade              bool
+		teamMode, singleMode bool
+	)
+
 	ha := &helmArgs{
 		cmdType: connector.HelmRequest_INSTALL,
 	}
@@ -41,6 +47,14 @@ func helmInstallCommand() *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if upgrade {
 				ha.cmdType = connector.HelmRequest_UPGRADE
+			}
+			switch {
+			case teamMode && singleMode:
+				return fmt.Errorf("flags `--team-mode` and `--single-mode` are mutually exclusive")
+			case teamMode:
+				ha.mode = manager.Mode_MODE_TEAM
+			case singleMode:
+				ha.mode = manager.Mode_MODE_SINGLE
 			}
 			return ha.run(cmd, args)
 		},
@@ -53,6 +67,8 @@ func helmInstallCommand() *cobra.Command {
 	flags.BoolVarP(&upgrade, "upgrade", "u", false, "replace the traffic manager if it already exists")
 	flags.StringSliceVarP(&ha.values, "values", "f", []string{}, "specify values in a YAML file or a URL (can specify multiple)")
 	flags.StringSliceVarP(&ha.valuePairs, "set", "", []string{}, "specify a value as a.b=v (can specify multiple or separate values with commas: a.b=v1,a.c=v2)")
+	flags.BoolVarP(&teamMode, "team-mode", "", false, "set the traffic-manager to team mode")
+	flags.BoolVarP(&singleMode, "single-mode", "", false, "set the traffic-manager to single user mode")
 
 	ha.request, ha.kubeFlags = initConnectRequest(cmd)
 	return cmd
@@ -101,6 +117,23 @@ func (ha *helmArgs) run(cmd *cobra.Command, _ []string) error {
 		ValuePairs:     ha.valuePairs,
 		ConnectRequest: ha.request,
 	}
+
+	if ha.mode != manager.Mode_MODE_UNSPECIFIED {
+		switch ha.mode {
+		case manager.Mode_MODE_SINGLE:
+			request.ValuePairs = append(request.ValuePairs, "trafficManager.mode=single")
+		case manager.Mode_MODE_TEAM:
+			request.ValuePairs = append(request.ValuePairs, "trafficManager.mode=team")
+		}
+
+		upgrade := ha.cmdType == connector.HelmRequest_UPGRADE
+		setFlagUsed := 0 < len(ha.valuePairs)
+		valuesFlagUsed := 0 < len(ha.values)
+		if upgrade && !setFlagUsed && !valuesFlagUsed {
+			request.ReuseValues = true
+		}
+	}
+
 	util.AddKubeconfigEnv(request.ConnectRequest)
 	resp, err := userD.Helm(ctx, request)
 	if err != nil {